requirements.txt
loguru==0.6.0
python-dotenv==0.21.1
python-telegram-bot==20.2
requests==2.28.2
pixivpy_async
psutil
